feat(phy/nb-iot): Refactor UL/DL PHY to use safe int32_t packed complex samples
This extensive refactoring migrates the core NB-IoT Physical Layer data path
from using unsafe pointer-casting over int32_t buffers (violating the Strict
Aliasing Rule) to a safe, packed int32_t representation where I and Q (or two
LLR components) are accessed via bitwise shifting. This change is mandatory
for compliance and enables future performance gains via SIMD vectorization.

Key Changes Applied Across Modules:

1.  **Data Typing and Allocation:**
    - Array declarations (e.g., LLR buffers, DFT temporaries) are consistently changed from  arrays/pointers to  arrays/pointers.
    - Memory allocation sizes (/              total        used        free      shared  buff/cache   available
Mem:       16422828      127828     8193524       15132     8101476    15816068
Swap:        998396           0      998396) are halved to correctly match the word count (), and padding is adjusted accordingly (e.g., from  to  words).

2.  **Coding and Transport Layers:**
    - **:** Function signature updated to receive . Implements safe unpacking (, ) within its core loop.
    - **:** Signature updated. Implements safe packing logic to write LLRs as  into the  output buffer.
    - **:** Updates local buffer sizing and function calls to match the new  signatures.

3.  **Modulation and Estimation (FEP/Channel Est.):**
    - **:** DFT function pointer () and local buffer () updated to . All unsafe pointer casts to  are removed in favor of direct  pointers. Requires updating external DFT/IDFT signatures.
    - **:** All local complex channel pointers (, , etc.) are converted to . All previous  access patterns are replaced with bitwise packing/unpacking, and auxiliary function calls (e.g., ) are updated to accept  pointers.

4.  **Interface Layer (IF4p5):**
    - **:** Pointer arithmetic is eliminated in both the  (DL) and  (UL) blocks. Data is safely extracted from / using bitwise shift/mask before A-law compression.

+++ openair1/PHY/INIT/lte_init.c
@@ -1606,7 +1606,7 @@ int phy_init_lte_eNB(PHY_VARS_eNB *eNB,
           }
         } //eNB_id
 
-        pusch_vars[UE_id]->llr = (int16_t*)malloc16_clear( (8*((3*8*6144)+12))*sizeof(int16_t) );
+        pusch_vars[UE_id]->llr = (int32_t*)malloc16_clear( (8*((3*8*6144)+12))*sizeof(int32_t) );
       } // abstraction_flag
     } //UE_id
