+++ openair1/PHY/LTE_TRANSPORT/extern_NB_IoT.h
@@ -36,8 +36,8 @@ extern int G_tab[18];
 
 
 extern short conjugate[8],conjugate2[8];
-extern short *ul_ref_sigs_rx_NB_IoT[30][4]; // NB-IoT: format 1 pilots
-extern short *ul_ref_sigs_f2_rx_NB_IoT[16]; // NB-IoT: format 2 pilots
+extern int32_t *ul_ref_sigs_rx_NB_IoT[30][4]; // NB-IoT: format 1 pilots
+extern int32_t *ul_ref_sigs_f2_rx_NB_IoT[16]; // NB-IoT: format 2 pilots
 extern unsigned short dftsizes[33];
